breaking up the generator module into various modules
